ai_foloup_hr_backend_v2-main/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic
redis==4.6.0
python-socketio
aiohttp
websockets
pydub
PyPDF2
python-docx
openai
pyyaml
python-multipart
email-validator
passlib
python-jose==3.5.0
bcrypt==4.3.0
boto3==1.28.39
pandas>=2.0.0
openpyxl>=3.0.0